Update EfaConfig.java
updateValue(): Fixed a bug which allowed config items starting with "_" to be stored in the file. This is not allowed (see comment in line 2657.

updateConfigValuesWithPersistence(): Fixed a bug when an item starting with "_" is found and the loop needs to ignore it - the iterator needed to be pushed to the next item in the list (otherwise endless loop when such an item is found).