PRETTY_BOATLISTS
EfaConfig.java
-----
Added a new expert type config parameter for activating the new look of the boatlists on the front frame of efaBoathouse. Default activated.

ItemTypeBoatstatusList.java
----
Constructor with new parameter to activate the new rendering.
Refactored the destination delimiter string to a constant.

EfaBoathouseFrame.java
-----
Each boat list (and person list) on the frame is initialized with the config parameter for better rendering.

ItemTypeList.java
----
New rendering.
- Separators are centered, grey background, no hyphens
- Boats on the Water/Not available boats:
  * boat name is aligned left-handed, the rest of the available space is used for the destination string. Destination string is truncated and extended with an ellipsis when there is not enough space.
Boat name and destination are distinguished by the specific separator of the listitems' text. This is a bit ugly, but currently there is no easy way to obtain the data otherwise.