Added Export/Import function for clubwork, fixed nullpointer if export ARGs were empty
MenuData.java
-------
removeOptionsFromArgs()
- can handle zero arguments and avoids nullpointer exception.

exportData()
- does not crash when user misses to specify a filename

printHelpContext()
- better help for actual usable encodings (although others can be specified)

CLI.java
--------
Added clubworkbase and clubworkboathouse as new menus

MenuBase.java
-------
printUsage()
- max length of a command/menu name can be 17 characters long ("clubworkboathouse"), so better alignment for

New Classes: MenuClubWorkBase and MenuClubWorkBoatHouse
- there can be diffrent clubwork data files for efaBase and efaBoatHouse.
So for each there is a single menu so that the user can choose between them by specifying the command. This is due to keeping clubwork exports compatible to all functinons in MenuData.java

MenuMain.java
-----------
- new menu items for clubworkBase and clubworkBoatHouse

Project.java
-------------
getCurrentClubworkForCLI()
Diffrent clubwork databases can be specified for efaBase and efaBoatHouse. The getCurrentClubWork decides on the APPID which one shall be opened. This does not suit for efaCLI. So new Function getCurrentClubworkForCLI() added.