Update, syncing with nmichael:efa:master
ItemTypeBoatStatusList.java
-------
- getSecondaryItem(), getBoatreservation()
instead of a constant value for the next hours of the day, we use the calculated remaining minutes for today (for getting today's reservations)

KanuEfbSyncTask.java
-------
- updated comments
- better logging of unknown boats
if boat is unknown, but at least one crew member has an efbID, the trip gets synced (for those members with an efbID), and this is logged also in non-verbose mode.

- syncUsers(), handleSyncUserResponse()
removed as no longer needed

BoatReservationListDialog.java
-----
- itemListenerAction()
the checkbox in the south part of the dialog "filtern" stays checked also if "search for today's reservations" is unchecked.

EfaBaseFrame.java
-----------
- added fix from nick for autocomplete lists permanently not showing crew members who had a new trip in the current running session of efaBase

EfaBoathouseFrame.java
------
- added some comments

AutoCompleteList.java
------
- added fix from nick for autocomplete lists permanently not showing crew members who had a new trip in the current running session of efaBase