Update efaBaseFrame.
fixed Check order in efabaseframe so that the checks are run in the order of the items on the screen.

if a boat is already on the water, the boat field gets the focus.

multisession mode: if a boat is not allowed to use, but the user decides to use it anyway, a log entry is created just as if we were in efaBaseFrame for a single trip.